requirements.txt
torch>=1.10.1
torchvision>=0.10.0
tqdm
numpy
matplotlib
pyyaml>=5.2.1
pytorch-transformers
prettytable
humanize